feat: add @appium/extdata module
This PR is a prerequisite to getting "npm-able" extensions to work.

This module provides a command-line tool and API to work with extension manifest (`extensions.yaml`) files, including info dumps and "migrations" from storing extensions in the _default_ `$APPIUM_HOME` to keeping them locally installed, relative to a project, and listed as deps within `package.json`.

In addition, the package provides tools to fix/sync an `extensions.yaml` with whatever's on disk, and an info dump about the system's Appium installation(s).

The gist of it is that if a project's `package.json` contains `appium`, then we will store the manifest file within `node_modules/.cache/appium/extensions.yaml`, relative to the project.  `$APPIUM_HOME` is then treated as the project root directory.  Or something along those lines, anyway--in a nutshell, the extensions and the manifest will all live in `node_modules` and the extensions will be "normal" dependencies, and the manifest becomes a compat layer.

The `appium` CLI has not yet been updated to support this behavior.

Additionally:

- broke the `tsconfig.json` up a bit because different modules need different things -- `@appium/extdata` is written in Node.js-compatible ESM and is not transpiled.
- provided an alternate, pass-through API for `klaw` in `@appium/support`, which enables use as an async iterator (frankly, much easier to deal with in the general case than the bespoke wrapper)
- fixed some context problems in `@appium/support`'s `hasAccess`, `exists`, `copyFile`, and `md5` methods; these were required to be called as methods on an `fs` context object for no discernable reason.